requirements.txt
# Backend
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.9
motor==3.7.0

# Visão Computacional
opencv-python==4.9.0.80
numpy==1.26.0
tensorflow==2.19.0

# Banco de Dados
pymongo
firebase-admin==6.2.0

# WebSockets
websockets==12.0
python-socketio==5.11.2

# Autenticação e Segurança
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
gdown==4.7.3

# Testes
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0